@@ -1,27 +1,27 @@
 import Configuration from './services/ConfigurationService';
-import { DroneState } from "./states/DroneState";
+import { DroneState } from './states/DroneState';
 import { ModulesManager } from './managers/ModulesManager';
 import { ModulesManagerBuilder } from './builders/ModulesManagerBuilder';
-import { DroneConfiguration, InitializationMode } from './interfaces/Configuration';
-let chalk: any = require('chalk');
+import { IDroneConfiguration, InitializationMode } from './interfaces/Configuration';
+import chalk from 'chalk';
 
 /**
  * BUILDING MODULES.
  */
-let state = new DroneState();
-let modulesManager = new ModulesManager(state);
+const state = new DroneState();
+const modulesManager = new ModulesManager(state);
 
-switch(Configuration.mode) {
-    case InitializationMode.NORMAL:
-        console.log(chalk.blue('Initialization mode: NORMAL'));
-        ModulesManagerBuilder.BuildHardwareModules(modulesManager);
-        break;
-    case InitializationMode.SIMULATION:
-        console.log(chalk.blue('Initialization mode: SIMULATION'));
-        ModulesManagerBuilder.BuildSimulationModules(modulesManager);
-        break;
-    default:
-        throw new Error('Invalid initialization mode.');
+switch (Configuration.mode) {
+	case InitializationMode.NORMAL:
+		console.log(chalk.blue('Initialization mode: NORMAL'));
+		ModulesManagerBuilder.BuildHardwareModules(modulesManager);
+		break;
+	case InitializationMode.SIMULATION:
+		console.log(chalk.blue('Initialization mode: SIMULATION'));
+		ModulesManagerBuilder.BuildSimulationModules(modulesManager);
+		break;
+	default:
+		throw new Error('Invalid initialization mode.');
 }
 
 ModulesManagerBuilder.BuildLogicModules(modulesManager);
@@ -29,18 +29,18 @@ ModulesManagerBuilder.BuildLogicModules(modulesManager);
 /**
  * STARTING.
  */
-if(process.argv[2] !== undefined && process.argv[2] === 'check') {
-    modulesManager.checkAll().then(() => {
-        process.exit();
-    }).then(() => {
-        process.exit();
-    });
+if (process.argv[2] !== undefined && process.argv[2] === 'check') {
+	modulesManager.checkAll().then(() => {
+		process.exit();
+	}).then(() => {
+		process.exit();
+	});
 } else {
-    modulesManager.enableAll().catch(() => {
-        console.log(chalk.red('System critical error on 1 module'));
-        modulesManager.disableAll();
-        process.exit();
-    });
+	modulesManager.enableAll().catch(() => {
+		console.log(chalk.red('System critical error on 1 module'));
+		modulesManager.disableAll();
+		process.exit();
+	});
 }
 
 /**
@@ -48,23 +48,24 @@ if(process.argv[2] !== undefined && process.argv[2] === 'check') {
  */
 process.stdin.setEncoding('utf8');
 process.stdin.on('readable', () => {
-    var chunk = process.stdin.read();
-    if (chunk !== null) {
-        switch(chunk.toString().trim()) {
-            case 'shutdown':
-            case 'exit':
-            case 'close':
-            case 'q':
-                console.log(chalk.yellow('Shutdown command received.'));
-                modulesManager.disableAll().then(() => {
-                    process.exit();
-                }, () => {
-                    process.exit();
-                });
-            break;
-            default:
-                console.log(chalk.grey('Command not found: ' + chunk));
-                break;
-        }
-    }
+	const chunk = process.stdin.read();
+	if (chunk !== null) {
+		switch (chunk.toString().trim()) {
+			case 'shutdown':
+			case 'exit':
+			case 'close':
+			case 'c':
+			case 'q':
+				console.log(chalk.yellow('Shutdown command received.'));
+				modulesManager.disableAll().then(() => {
+					process.exit();
+				}, () => {
+					process.exit();
+				});
+				break;
+			default:
+				console.log(chalk.grey('Command not found: ' + chunk));
+				break;
+		}
+	}
 });
