@@ -1,147 +1,151 @@
-import { DroneModule } from '../interfaces/Module'
-import { DroneState } from "../states/DroneState";
-let chalk: any = require('chalk');
+import { IDroneModule } from '../interfaces/Module';
+import { DroneState } from '../states/DroneState';
+import chalk from 'chalk';
 
 export class ModulesManager {
-    private state: DroneState;
-    private modules: Array<DroneModule> = [];
-    private moduleInitializationTimeout = 10000;
+	private state: DroneState;
+	private modules: Array<IDroneModule> = [];
+	private moduleInitializationTimeout = 10000;
 
-    constructor(state: DroneState) {
-        this.state = state;
-    }
+	constructor(state: DroneState) {
+		this.state = state;
+	}
 
-    public loadModule(modulePath: string) {
-        let moduleName = modulePath.split('/').pop();
-        let moduleLib = require('../modules/' + modulePath);
-        let moduleInstance = new moduleLib[moduleName]();
-        this.add(moduleInstance);
-    }
+	public loadModule(modulePath: string) {
+		const moduleName = modulePath.split('/').pop();
+		const moduleLib = require('../modules/' + modulePath);
+		const moduleInstance = new moduleLib[moduleName]();
+		this.add(moduleInstance);
+	}
 
-    public add(m: DroneModule) {
-        m.setState(this.state);
-        this.modules.push(m);
-    }
+	public add(m: IDroneModule) {
+		m.setState(this.state);
+		this.modules.push(m);
+	}
 
-    public checkAll(): Promise<string> {
-        console.log(chalk.cyan("Checking sequence started."));
+	public checkAll(): Promise<string> {
+		console.log(chalk.cyan('Checking sequence started.'));
 
-        return new Promise<string>((resolve, reject) => {
-            var onShutdownEnd = () => {
-                console.log(chalk.cyan("Shutdown checking complete."));
-                console.log(chalk.cyan("Checking sequence finished."));
-            };
+		return new Promise<string>((resolve, reject) => {
+			const onShutdownEnd = () => {
+				console.log(chalk.cyan('Shutdown checking complete.'));
+				console.log(chalk.cyan('Checking sequence finished.'));
+			};
 
-            var onInitEnd = () => {
-                console.log(chalk.cyan("Initialization checking complete."));
-                console.log(chalk.cyan("Systems will shutdown in 3 seconds."));
+			const onInitEnd = () => {
+				console.log(chalk.cyan('Initialization checking complete.'));
+				console.log(chalk.cyan('Systems will shutdown in 3 seconds.'));
 
-                setTimeout(() => {
-                    this.disableAll().then(() => {
-                        onShutdownEnd();
-                        resolve();
-                    }, () => {
-                        onShutdownEnd();
-                        reject();
-                    });
-                }, 3000);
-            };
+				setTimeout(() => {
+					this.disableAll().then(() => {
+						onShutdownEnd();
+						resolve();
+					}, () => {
+						onShutdownEnd();
+						reject();
+					});
+				}, 3000);
+			};
 
-            this.enableAll().then(() => {
-                onInitEnd();
-            }, () => {
-                onInitEnd();
-            });
-        });
-    }
+			this.enableAll().then(() => {
+				onInitEnd();
+			}, () => {
+				onInitEnd();
+			});
+		});
+	}
 
-    public enableAll(): Promise<string> {
-        console.log(chalk.inverse("Initializing Modules"));
+	public logStatus(status: string, message: string) {
+		console.log(chalk.white('[ ') + status + chalk.white(' ] ') + chalk.white(message));
+	}
 
-        return new Promise<string>((resolve, reject) => {
-            let i = 0;
-            let fails = 0;
+	public enableAll(): Promise<string> {
+		console.log(chalk.inverse('Initializing Modules'));
 
-            let onSuccessfull = (m: DroneModule) => {
-                console.log('[ ' + chalk.green("DONE") + ' ] ' + m.name);
-            };
+		return new Promise<string>((resolve, reject) => {
+			let i = 0;
+			let fails = 0;
 
-            let onFail = (m: DroneModule, message: string) => {
-                fails++;
-                console.log('[ ' + chalk.red("FAIL") + ' ] ' + m.name);
-                console.log(chalk.magenta(message));
-            };
+			const onSuccessfull = (m: IDroneModule) => {
+				this.logStatus(chalk.green('DONE'), m.name);
+			};
 
-            let onEnd = () => {
-                if(fails === 0) {
-                    console.log('[ ' + chalk.green("ONLINE") + ' ] ' + 'All modules successfully initialized.');
-                    resolve();
-                } else {
-                    console.log('[ ' + chalk.red("ERROR") + ' ] ' + fails + ' modules failed during initialization.');
-                    reject();
-                }
-            };
+			const onFail = (m: IDroneModule, message: string) => {
+				fails++;
+				this.logStatus(chalk.red('FAIL'), m.name);
+				console.log(chalk.magenta(message));
+			};
 
-            let next = () => {
-                i++;
-                if(i < this.modules.length) {
-                    checkModule();
-                } else {
-                    onEnd();
-                }
-            };
+			const onEnd = () => {
+				if (fails === 0) {
+					this.logStatus(chalk.blue('ONLINE'), 'All modules successfully initialized.');
+					resolve();
+				} else {
+					this.logStatus(chalk.red('ERROR'), fails + ' modules failed during initialization.');
+					reject();
+				}
+			};
 
-            let checkModule = () => {
-                let m = this.modules[i];
+			const next = () => {
+				i++;
+				if (i < this.modules.length) {
+					checkModule();
+				} else {
+					onEnd();
+				}
+			};
 
-                let timeout = setTimeout(() => {
-                    onFail(m, 'Timeout - No response.');
-                    next();
-                    timeout = null;
-                }, this.moduleInitializationTimeout);
+			const checkModule = () => {
+				const m = this.modules[i];
 
-                m.enable().then(() => {
-                    if(timeout !== null) {
-                        onSuccessfull(m);
-                        next();
-                        clearTimeout(timeout);
-                        timeout = null;
-                    }
-                }, (message: string) => {
-                    if(timeout !== null) {
-                        onFail(m, message);
-                        next();
-                        clearTimeout(timeout);
-                        timeout = null;
-                    }
-                });
+				let timeout = setTimeout(() => {
+					onFail(m, 'Timeout - No response.');
+					next();
+					timeout = null;
+				}, this.moduleInitializationTimeout);
 
-            };
-            
-            checkModule();
-        });
-    }
+				m.enable().then(() => {
+					if (timeout !== null) {
+						onSuccessfull(m);
+						next();
+						clearTimeout(timeout);
+						timeout = null;
+					}
+				}, (message: string) => {
+					if (timeout !== null) {
+						onFail(m, message);
+						next();
+						clearTimeout(timeout);
+						timeout = null;
+					}
+				});
 
-    public disableAll(): Promise<any> {
-        console.log(chalk.inverse("Shutting down Modules"));
+			};
 
-        let promises: Array<Promise<string>> = [];
+			checkModule();
+		});
+	}
 
-        for(let i = (this.modules.length-1); i >= 0; i--) {
-            let promise = this.modules[i].disable();
-            promise.catch((message) => {
-                console.log('[ ' + chalk.red("ERROR") + ' ] ' + this.modules[i].name);
-                console.log(chalk.magenta(message));
-            });
-            promises.push(promise);
-        }
+	public disableAll(): Promise<any> {
+		console.log(chalk.inverse('Shutting down Modules'));
 
-        console.log(chalk.grey("Shutdown command was sent to all modules."));
+		const promises: Array<Promise<string>> = [];
 
-        return Promise.all(promises).then(() => {
-            console.log('[ ' + chalk.grey("SHUTDOWN") + ' ] ' + 'All modules successfully disabled.');
-        }, () => {
-            console.log('[ ' + chalk.red("SHUTDOWN FAILED") + ' ] ' + 'Some modules failed during shutdown.');
-        });
-    }
+		for (let i = (this.modules.length - 1); i >= 0; i--) {
+			const promise = this.modules[i].disable();
+			promise.catch(message => {
+				this.logStatus(chalk.red('ERROR'), this.modules[i].name);
+				console.log(chalk.magenta(message));
+			});
+			promises.push(promise);
+		}
+
+		console.log(chalk.grey('Shutdown command was sent to all modules.'));
+
+		return Promise.all(promises).then(() => {
+			this.logStatus(chalk.grey('SHUTDOWN'), 'All modules successfully disabled.');
+		}, () => {
+			this.logStatus(chalk.red('SHUTDOWN FAILED'), 'Some modules has failed during shutdown.');
+		});
+	}
 }
