@@ -1,20 +1,15 @@
-/// <reference path="../../../typings/globals/cannon/index.d.ts" />
-import { DroneState } from "../../states/DroneState";
-import { DroneModule } from '../../interfaces/Module';
+import { DroneState } from '../../states/DroneState';
+import { IDroneModule } from '../../interfaces/Module';
 import { Utils } from '../../helpers/Utils';
-import BoardService from "../../services/BoardService";
+import BoardService from '../../services/BoardService';
 import Configuration from '../../services/ConfigurationService';
-let CANNON = require('cannon');
+import * as CANNON from 'cannon';
 
-export class IMU implements DroneModule {
+export class IMU implements IDroneModule {
 	public name: string = 'Inertial Measurement Unit (IMU)';
 	private state: DroneState;
 	private disposers: Array<any> = [];
 	private imu: any;
-
-	constructor () {
-
-	}
 
 	public setState(state: DroneState) {
 		this.state = state;
@@ -22,25 +17,25 @@ export class IMU implements DroneModule {
 
 	public enable(): Promise<string> {
 		return new Promise<string>((resolve, reject) => {
-			BoardService.getBoard(Configuration.imu.board).then((board) => {
-				let five: any = require("johnny-five");
+			BoardService.getBoard(Configuration.imu.board).then(board => {
+				const five: any = require('johnny-five');
 				this.imu = new five.IMU({
-					board: board,
-					controller: "BNO055",
+					board,
+					controller: 'BNO055',
 					enableExternalCrystal: false,
 					calibrationMask: 48
 				});
 
 				this.configureActions();
 
-				this.imu.on("calibration", (calibration: any) => {
+				this.imu.on('calibration', (calibration: any) => {
 					resolve();
-					if(calibration == 179) {
+					if (calibration === 179) {
 						console.log('IMU calibration complete!');
 						this.state.current.calibratedImu.setValue(true);
 					}
 				});
-			});			
+			});
 		});
 	}
 
@@ -54,9 +49,9 @@ export class IMU implements DroneModule {
 	}
 
 	private configureActions() {
-		var state = this.state;
+		const state = this.state;
 
-		this.imu.on("change", function(err:any, data:any) {
+		this.imu.on('change', function(err: any, data: any) {
 			state.current.accelerometer.setValue({
 				// TODO: re-map vector 90 deg CW
 				vector: new CANNON.Vec3(this.accelerometer.x, this.accelerometer.y, this.accelerometer.z),
@@ -66,15 +61,15 @@ export class IMU implements DroneModule {
 			state.current.temperature.setValue(this.thermometer.celsius);
 		});
 
-		this.imu.orientation.on("change", function(err:any, data:any) {
+		this.imu.orientation.on('change', function(err: any, data: any) {
 			// TODO: re-map quaternion 90 deg CW
 			try {
 				let q = new CANNON.Quaternion();
-				let rot = new CANNON.Quaternion();
+				const rot = new CANNON.Quaternion();
 
-				rot.set(this.quarternion.x,this.quarternion.y,this.quarternion.z,this.quarternion.w);
+				rot.set(this.quarternion.x, this.quarternion.y, this.quarternion.z, this.quarternion.w);
 				q = q.mult(rot);
-				//q = q.inverse();
+				// q = q.inverse();
 
 				// rot.setFromAxisAngle(new CANNON.Vec3(0,1,0), Utils.toRadians(-90));
 				// q = q.mult(rot);
@@ -83,7 +78,7 @@ export class IMU implements DroneModule {
 				// q = q.mult(rot);
 
 				state.current.orientation.setValue(rot);
-			} catch(e) {
+			} catch (e) {
 				console.log('warning error!');
 				console.log(e);
 			}
