@@ -1,16 +1,15 @@
-/// <reference path="../../typings/globals/cannon/index.d.ts" />
-import { StateProperty } from "../helpers/StateProperty";
-let CANNON = require('cannon');
+import { StateProperty } from '../helpers/StateProperty';
+import * as CANNON from 'cannon';
 
-interface AccelerometerStateValue {
-	vector: CANNON.Vec3,
-	acceleration: Number
+interface IAccelerometerStateValue {
+	vector: CANNON.Vec3;
+	acceleration: number;
 }
 
-export class AccelerometerState extends StateProperty<AccelerometerStateValue> {
-	constructor () {
+export class AccelerometerState extends StateProperty<IAccelerometerStateValue> {
+	constructor() {
 		super({
-			vector: new CANNON.Vec3(0,0,0),
+			vector: new CANNON.Vec3(0, 0, 0),
 			acceleration: 0
 		});
 	}
