@@ -1,9 +1,8 @@
-/// <reference path="../../typings/globals/cannon/index.d.ts" />
-import { StateProperty } from "../helpers/StateProperty";
-let CANNON = require('cannon');
+import { StateProperty } from '../helpers/StateProperty';
+import * as CANNON from 'cannon';
 
 export class GyroscopeState extends StateProperty<CANNON.Vec3> {
-	constructor () {
-		super(new CANNON.Vec3(0,0,0));
+	constructor() {
+		super(new CANNON.Vec3(0, 0, 0));
 	}
 }
