@@ -1,9 +1,8 @@
-/// <reference path="../../typings/globals/cannon/index.d.ts" />
-import { StateProperty } from "../helpers/StateProperty";
-let CANNON = require('cannon');
+import { StateProperty } from '../helpers/StateProperty';
+import * as CANNON from 'cannon';
 
 export class OrientationState extends StateProperty<CANNON.Quaternion> {
-	constructor () {
-		super(new CANNON.Quaternion(0,0,0,1));
+	constructor() {
+		super(new CANNON.Quaternion(0, 0, 0, 1));
 	}
 }
