@@ -1,18 +1,14 @@
-import { DroneState } from "../../states/DroneState";
-import { DroneModule } from '../../interfaces/Module';
+import { DroneState } from '../../states/DroneState';
+import { IDroneModule } from '../../interfaces/Module';
 import Configuration from '../../services/ConfigurationService';
 
-export class DCSensor implements DroneModule {
+export class DCSensor implements IDroneModule {
 	public name: string = 'DC Sensor';
 	private state: DroneState;
 	private disposers: Array<any> = [];
 
 	private fakeInitValue = Configuration.battery.voltageMax;
 	private startTime: number;
-
-	constructor () {
-		
-	}
 
 	public setState(state: DroneState) {
 		this.state = state;
@@ -37,16 +33,16 @@ export class DCSensor implements DroneModule {
 	private configureActions() {
 		this.startTime = Date.now();
 
-        this.disposers.push(this.state.simulation.position
-            .getStream()
-            .onValue((enginesState) => {
-				let elapsed = Date.now() - this.startTime;
-				let dischargeRage = 0.00001; // fake, for simulation purposes.
-				let current = this.fakeInitValue - (elapsed * dischargeRage);
+		this.disposers.push(this.state.simulation.position
+			.getStream()
+			.onValue(enginesState => {
+				const elapsed = Date.now() - this.startTime;
+				const dischargeRage = 0.00001; // fake, for simulation purposes.
+				const current = this.fakeInitValue - (elapsed * dischargeRage);
 
-                this.state.current.dcSensor.setValue({
+				this.state.current.dcSensor.setValue({
 					voltage: current
 				});
-            }));
+			}));
 	}
 }
