@@ -1,23 +1,24 @@
 import { Utils } from '../helpers/Utils';
-import { DroneConfiguration } from '../interfaces/Configuration'
+import { IDroneConfiguration } from '../interfaces/Configuration';
 
 class ConfigurationService {
-	private static instance: ConfigurationService = null;
-    private settings: DroneConfiguration;
+	public static getSettings(): IDroneConfiguration {
+		if (ConfigurationService.instance === null) {
+			ConfigurationService.instance = new ConfigurationService();
+		}
 
-	private constructor () {
-        let defaults: DroneConfiguration = require('../../config/default.config.json');
-        let env = require('../../config/env.config.json');
-        this.settings = defaults;
-        Utils.deepExtend(this.settings, env);
+		return ConfigurationService.instance.settings;
 	}
 
-    public static getSettings():DroneConfiguration {
-        if(ConfigurationService.instance === null)
-            ConfigurationService.instance = new ConfigurationService();
-        
-        return ConfigurationService.instance.settings;
-    }
+	private static instance: ConfigurationService = null;
+	private settings: IDroneConfiguration;
+
+	private constructor() {
+		const defaults: IDroneConfiguration = require('../../config/default.config.json');
+		const env = require('../../config/env.config.json');
+		this.settings = defaults;
+		Utils.deepExtend(this.settings, env);
+	}
 }
 
 export default ConfigurationService.getSettings();
