@@ -1,9 +1,9 @@
-import { StateProperty } from "../helpers/StateProperty";
+import { StateProperty } from '../helpers/StateProperty';
 
 export class CommunicationState {
-    public connected: StateProperty<boolean>;
+	public connected: StateProperty<boolean>;
 
-	constructor () {
-		this.connected = new StateProperty<boolean>(false); 
+	constructor() {
+		this.connected = new StateProperty<boolean>(false);
 	}
 }
