@@ -1,14 +1,14 @@
-import { StateProperty } from "../helpers/StateProperty";
+import { StateProperty } from '../helpers/StateProperty';
 
-interface SteeringValue {
+export interface ISteeringValue {
 	yaw: number;
 	pitch: number;
 	roll: number;
 	throttle: number;
 }
 
-export class SteeringState extends StateProperty<SteeringValue> {
-	constructor () {
+export class SteeringState extends StateProperty<ISteeringValue> {
+	constructor() {
 		super({
 			yaw: 0,
 			pitch: 0,
